.저장하기위한커밋

+++ plse/src/main/java/com/care/dto/BoardDTO.java
@@ -13,7 +13,7 @@ public class BoardDTO {
 	private int hit;
 	private int recommend;
 	private Date savedate;
-	private int stars;
+	private double stars;
 	private String productname;
 	private String content;
 	private int cvsnum;
